% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{abs_posit32}
\alias{abs_posit32}
\title{abs}
\usage{
\S4method{abs}{posit32}(x)
}
\arguments{
\item{x}{posit32 object.}
}
\value{
A posit32 object.
}
\description{
Returns the absolute value of posit32 object.This is to override the default abs function.
}
\examples{
a = new ("posit32" , v = -3L)
b = abs(a)
}
