% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{abs_posit8}
\alias{abs_posit8}
\title{abs}
\usage{
\S4method{abs}{posit8}(x)
}
\arguments{
\item{x}{posit8 object.}
}
\value{
A posit8 object.
}
\description{
Returns the absolute value of posit8 object.This is to override the default abs function.
}
\examples{
a = new ("posit8" , v = -3L)
b = abs(a)
}
