% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{abs_posit_2}
\alias{abs_posit_2}
\title{abs}
\usage{
\S4method{abs}{posit_2}(x)
}
\arguments{
\item{x}{posit_2 object.}
}
\value{
A posit_2 object.
}
\description{
Returns the absolute value of posit_2 object.This is to override the default abs function.
}
\examples{
a = new ("posit_2" , v = -3L , y = 5L)
b = abs(a)
}
