% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{AND}
\alias{AND}
\title{AND}
\usage{
AND(x, y)
}
\arguments{
\item{x}{A posit8, posit16, posit32 or posit_2 object.}

\item{y}{A posit8, posit16, posit32 or posit_2 object.}
}
\value{
posit8, posit16, posit32 or posit_2 object.
}
\description{
Performs AND operation on two same posit object types.
}
