% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{clr}
\alias{clr}
\title{clr}
\usage{
clr(x)
}
\arguments{
\item{x}{A quire8, quire16, quire32 or quire_2 object.}
}
\value{
quire8, quire16, quire32 or quire_2 object.
}
\description{
Clears quire object.
}
\examples{
a = new ("quire8")
b = clr(a)
a = new ("quire16")
b = clr(a)
a = new ("quire32")
b = clr(a)
a = new ("quire_2" , y = 6L)
b = clr(a)
}
