% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{convertToColor}
\alias{convertToColor}
\title{Add together two numbers.}
\usage{
convertToColor(i, ps, es)
}
\arguments{
\item{i}{A number.}

\item{ps}{A number.}

\item{es}{A number.}
}
\value{
The sum of \code{i} and \code{ps}.
}
\description{
Add together two numbers.
}
