% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{eq_posit8}
\alias{eq_posit8}
\title{==}
\usage{
\S4method{==}{posit8,posit8}(e1, e2)
}
\arguments{
\item{e1}{posit8 object.}

\item{e2}{posit8 object.}
}
\value{
TRUE or FALSE.
}
\description{
Compares two posit8 type objects.This is to override the default == operator.
}
\examples{
a = new ("posit8" , v = 3L)
b = new ("posit8" , v = 4L)
a == b
}
