% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{fma}
\alias{fma}
\title{fma}
\usage{
fma(x, y, z)
}
\arguments{
\item{x}{A posit8, posit16, posit32 or posit_2 object.}

\item{y}{A posit8, posit16, posit32 or posit_2 object.}

\item{z}{A posit8, posit16, posit32 or posit_2 object.}
}
\value{
posit8, posit16, posit32 or posit_2 object.
}
\description{
Performs fused multiply-add between posit type objects.
}
\examples{
a = new ("posit8" , v = 1L)
b = new ("posit8" , v = 2L)
c = new ("posit8" , v = 3L)
d = fma (a,b,c)
}
