% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{init_posit16}
\alias{init_posit16}
\title{new}
\usage{
\S4method{initialize}{posit16}(.Object, ...)
}
\arguments{
\item{v}{An integer or a double value.}
}
\value{
posit16 object.
}
\description{
Initializes a posit16 object.
}
\examples{
a = new ("posit16" , v = 3L)
a = new ("posit16" , v = 3)
}
