% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{init_posit32}
\alias{init_posit32}
\title{new}
\usage{
\S4method{initialize}{posit32}(.Object, ...)
}
\arguments{
\item{v}{An integer or a double value.}
}
\value{
posit32 object.
}
\description{
Initializes a posit32 object.
}
\examples{
a = new ("posit32" , v = 30L)
a = new ("posit32" , v = 30)
}
