% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{init_posit8}
\alias{init_posit8}
\title{new}
\usage{
\S4method{initialize}{posit8}(.Object, ...)
}
\arguments{
\item{v}{An integer or a double value.}
}
\value{
posit8 object.
}
\description{
Initializes a posit8 object.
}
\examples{
a = new ("posit8" , v = 3L)
a = new ("posit8" , v = 3)
}
