% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{init_posit_2}
\alias{init_posit_2}
\title{new}
\usage{
\S4method{initialize}{posit_2}(.Object, ...)
}
\arguments{
\item{v}{An integer or a double value.}

\item{y}{An integer value representing the size of posit.}
}
\value{
posit_2 object.
}
\description{
Initializes a posit_2 object.
}
\examples{
a = new ("posit_2" , v = 30L , y = 16L)
a = new ("posit_2" , v = 30 , y = 16L)
}
