% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{init_quire16}
\alias{init_quire16}
\title{new}
\usage{
\S4method{initialize}{quire16}(.Object, ...)
}
\value{
quire16 object.
}
\description{
Initializes a quire16 object.
}
\examples{
a = new ("quire16")
}
