% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{init_quire32}
\alias{init_quire32}
\title{new}
\usage{
\S4method{initialize}{quire32}(.Object, ...)
}
\value{
quire32 object.
}
\description{
Initializes a quire32 object.
}
\examples{
a = new ("quire32")

}
