% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{init_quire8}
\alias{init_quire8}
\title{new}
\usage{
\S4method{initialize}{quire8}(.Object, ...)
}
\value{
quire8 object.
}
\description{
Initializes a quire8 object.
}
\examples{
a = new ("quire8")
}
