% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{init_quire_2}
\alias{init_quire_2}
\title{new}
\usage{
\S4method{initialize}{quire_2}(.Object, ...)
}
\arguments{
\item{y}{size of quire}
}
\value{
quire_2 object.
}
\description{
Initializes a quire_2 object.
}
\examples{
a = new ("quire_2", y = 4L)
}
