% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{isNaR}
\alias{isNaR}
\title{isNaR}
\usage{
isNaR(x)
}
\arguments{
\item{x}{A posit8, quire8, posit16, quire16, posit32, quire32, posit_2 or quire_2 object.}
}
\value{
True or False
}
\description{
Posit has only one bit pattern NaR (Not a Real) to represent infinity,
undefined and unrepresented values.isNaR returns true or false on the basis of input.
}
\examples{
a = new ("posit8" , v = 3L)
b = isNaR(a)
}
