% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{lshift}
\alias{lshift}
\title{lshift}
\usage{
lshift(x, y)
}
\arguments{
\item{x}{A posit8, posit16 or  posit32 object.}

\item{y}{An integer value representing the times the object should be shifted left.}
}
\value{
posit8, posit16 or posit32 object.
}
\description{
Left shifts posit8, posit16 or posit32 object by argument y.
}
\examples{
a = new ("posit8" , v = 3L)
b = lshift(a,5L)
a = new ("posit16" , v = 3L)
b = lshift(a,5L)
a = new ("posit32" , v = 3L)
b = lshift(a,5L)
a = new ("posit_2" , v = 4L, y=4L)
b = lshift(a,2L)
}
