% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{NOT}
\alias{NOT}
\title{NOT}
\usage{
NOT(x)
}
\arguments{
\item{x}{A posit8, posit16, posit32 or posit_2 object.}
}
\value{
posit8, posit16, posit32 or posit_2 object.
}
\description{
Performs invert (NOT) operation on posit object types.
}
\examples{
a = new ("posit8" , v = 3L)
c = NOT(a)
a = new ("posit16" , v = 3L)
c = NOT(a)
a = new ("posit32" , v = 3L)
c = NOT(a)
a = new ("posit_2", v =4L, y=5L)
c = NOT(a)
}
