% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{class}
\name{posit16-class}
\alias{posit16-class}
\title{posit16}
\description{
An S4 class to represent posit16
}
\section{Slots}{

\describe{
\item{\code{v}}{An input value to be converted to posit16}
}}

\examples{
x = new ("posit16" , v = 4L)
y = new ("posit16" , v = 5)
}
