% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{class}
\name{posit32-class}
\alias{posit32-class}
\title{posit32}
\description{
An S4 class to represent posit32
}
\section{Slots}{

\describe{
\item{\code{v}}{An input value to be converted to posit32}
}}

\examples{
x = new ("posit32" , v = 4L)
y = new ("posit32" , v = 5)
}
