% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{class}
\name{posit8-class}
\alias{posit8-class}
\title{posit8}
\description{
An S4 class to represent posit8
}
\section{Slots}{

\describe{
\item{\code{v}}{An input value to be converted to posit8}
}}

\examples{
x = new ("posit8" , v = 4L)
y = new ("posit8" , v = 5)
}
