% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{class}
\name{posit_2-class}
\alias{posit_2-class}
\title{posit_2}
\description{
An S4 class to represent posit_2
}
\section{Slots}{

\describe{
\item{\code{v}}{An input value to be converted to posit_2}

\item{\code{y}}{An integer value representing the size of posit}
}}

\examples{
a = new ("posit_2" , v = 4L , y = 6L)
b = new ("posit_2" , v = 5 , y = 8L)
}
