% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{qms}
\alias{qms}
\title{qms}
\usage{
qms(x, y, z)
}
\arguments{
\item{x}{A quire8, quire16, quire32 or quire_2 object.}

\item{y}{A posit8, posit16, posit32 or posit_2 object.}

\item{z}{A posit8, posit16, posit32 or posit_2 object.}
}
\value{
quire8, quire16, quire32 or quire_2 object.
}
\description{
Fused operation (qms) between quire and posit type object.
}
\examples{
a = new ("quire8")
b = new ("posit8", v = 4L)
c = new ("posit8", v = 5L)
a = qms (a,b,c)
}
