% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{class}
\name{quire_2-class}
\alias{quire_2-class}
\title{quire_2}
\description{
An S4 class to represent quire_2
}
\section{Slots}{

\describe{
\item{\code{y}}{An integer value representing the size of quire}
}}

\examples{
a = new ("quire_2" , y = 6L)
b = new ("quire_2" , y = 8L)
}
