% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{rint}
\alias{rint}
\title{rint}
\usage{
rint(x)
}
\arguments{
\item{x}{A posit8, posit16, posit32 or posit_2 object.}
}
\value{
A posit8, posit16, posit32 or posit_2 object.
}
\description{
Rounds to integer in posit type object
}
\examples{
a = new ("posit8" , v = 3.3)
b = rint(a)
}
