% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{rshift}
\alias{rshift}
\title{rshift}
\usage{
rshift(x, y)
}
\arguments{
\item{x}{A posit8, posit16 or  posit32 object.}

\item{y}{An integer value representing the times the object should be shifted right.}
}
\value{
posit8, posit16 or posit32 object.
}
\description{
Right shifts posit8, posit16 or posit32 object by argument y.
}
\examples{
a = new ("posit8" , v = 3L)
b = rshift(a,5L)
a = new ("posit16" , v = 3L)
b = rshift(a,5L)
a = new ("posit32" , v = 3L)
b = rshift(a,5L)
a = new ("posit_2" , v = 3L, y = 4L)
b = rshift(a,1L)
}
