% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{show_posit16}
\alias{show_posit16}
\title{show}
\usage{
\S4method{show}{posit16}(object)
}
\arguments{
\item{object}{posit16 object.}
}
\value{
posit16 object.
}
\description{
Returns the value of posit16 object on console.This method is invoked whether
we call print , show or simply write a posit16 object and hit enter .
}
\examples{
a = new ("posit16" , v = 3L)
a
show(a)
print(a)
}
