% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{show_posit32}
\alias{show_posit32}
\title{show}
\usage{
\S4method{show}{posit32}(object)
}
\arguments{
\item{object}{posit32 object.}
}
\value{
posit32 object.
}
\description{
Returns the value of posit32 object on console.This method is invoked whether
we call print , show or simply write a posit32 object and hit enter .
}
\examples{
a = new ("posit32" , v = 3L)
a
show(a)
print(a)
}
