% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{show_posit8}
\alias{show_posit8}
\title{show}
\usage{
\S4method{show}{posit8}(object)
}
\arguments{
\item{object}{posit8 object.}
}
\value{
posit8 object.
}
\description{
Returns the value of posit8 object on console.This method is invoked whether
we call print , show or simply write a posit8 object and hit enter .
}
\examples{
a = new ("posit8" , v = 3L)
a
show(a)
print(a)
}
