% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{show_posit_2}
\alias{show_posit_2}
\title{show}
\usage{
\S4method{show}{posit_2}(object)
}
\arguments{
\item{object}{posit_2 object.}
}
\value{
posit_2 object.
}
\description{
Returns the value of posit_2 object on console.This method is invoked whether
we call print , show or simply write a posit_2 object and hit enter .
}
\examples{
a = new ("posit_2" , v = 3L , y = 5L)
a
show(a)
print(a)
}
