% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{show_quire16}
\alias{show_quire16}
\title{show}
\usage{
\S4method{show}{quire16}(object)
}
\arguments{
\item{object}{quire16 object.}
}
\value{
quire16 object.
}
\description{
Returns the value of quire16 object on console.This method is invoked whether
we call print , show or simply write a quire16 object and hit enter .
}
\examples{
a = new ("quire16")
a
show(a)
print(a)

}
