% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{show_quire32}
\alias{show_quire32}
\title{show}
\usage{
\S4method{show}{quire32}(object)
}
\arguments{
\item{object}{quire32 object.}
}
\value{
quire32 object.
}
\description{
Returns the value of quire32 object on console.This method is invoked whether
we call print , show or simply write a quire32 object and hit enter .
}
\examples{
a = new ("quire32")
a
show(a)
print(a)

}
