% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{show_quire8}
\alias{show_quire8}
\title{show}
\usage{
\S4method{show}{quire8}(object)
}
\arguments{
\item{object}{quire8 object.}
}
\value{
quire8 object.
}
\description{
Returns the value of quire8 object on console.This method is invoked whether
we call print , show or simply write a quire8 object and hit enter .
}
\examples{
a = new ("quire8")
a
show(a)
print(a)

}
