% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{show_quire_2}
\alias{show_quire_2}
\title{show}
\usage{
\S4method{show}{quire_2}(object)
}
\arguments{
\item{object}{quire_2 object.}
}
\value{
quire_2 object.
}
\description{
Returns the value of quire_2 object on console.This method is invoked whether
we call print , show or simply write a quire_2 object and hit enter .
}
\examples{
a = new ("quire_2" , y = 2L)
a
show(a)
print(a)

}
