% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{sqrt_posit32}
\alias{sqrt_posit32}
\title{sqrt}
\usage{
\S4method{sqrt}{posit32}(x)
}
\arguments{
\item{x}{posit32 object.}
}
\value{
A posit32 object.
}
\description{
Returns the square root posit32 object.This is to override the default sqrt function.
}
\examples{
a = new ("posit32" , v = 3L)
b = sqrt(a)
}
