% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{sqrt_posit8}
\alias{sqrt_posit8}
\title{sqrt}
\usage{
\S4method{sqrt}{posit8}(x)
}
\arguments{
\item{x}{posit8 object.}
}
\value{
A posit8 object.
}
\description{
Returns the square root posit8 object.This is to override the default sqrt function.
}
\examples{
a = new ("posit8" , v = 3L)
b = sqrt(a)
}
