% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\docType{methods}
\name{sqrt_posit_2}
\alias{sqrt_posit_2}
\title{sqrt}
\usage{
\S4method{sqrt}{posit_2}(x)
}
\arguments{
\item{x}{posit_2 object.}
}
\value{
A posit_2 object.
}
\description{
Returns the square root posit_2 object.This is to override the default sqrt function.
}
\examples{
a = new ("posit_2" , v = 3L , y = 5L)
b = sqrt(a)
}
