% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{toBinary}
\alias{toBinary}
\title{toBinary}
\usage{
toBinary(x)
}
\arguments{
\item{x}{A posit8, quire8, posit16, quire16, posit32, quire32, posit_2 or quire_2 object.}
}
\value{
posit8, quire8, posit16, quire16, posit32, quire32, posit_2 or quire_2 output in binary format.
}
\description{
Converts posit or quire type object to binary format.
}
\examples{
a = new ("posit8" , v = 3L)
toBinary(a)
}
