% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{toBinaryFormatted}
\alias{toBinaryFormatted}
\title{toBinaryFormatted}
\usage{
toBinaryFormatted(x)
}
\arguments{
\item{x}{A posit8, posit16, posit32 or posit_2 object.}
}
\value{
posit8, posit16 or posit32 output in binary format.
}
\description{
Converts posit type object to binary format along with
color coding.Sign bit are color coded with red , regime bits with yellow, exponent bits with blue
and fraction bits with white
}
