% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{toDouble}
\alias{toDouble}
\title{toDouble}
\usage{
toDouble(x)
}
\arguments{
\item{x}{A posit8, posit16, posit32 or posit_2 object.}
}
\value{
output in double.
}
\description{
Converts posit type object to type double .
}
\examples{
a = new ("posit8" , v = 4L)
b = toDouble(a)
}
