% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{toHex}
\alias{toHex}
\title{toHex}
\usage{
toHex(x)
}
\arguments{
\item{x}{A posit8, quire8, posit16, quire16, posit32, quire32, posit_2 or quire_2 object.}
}
\value{
posit8, quire8, posit16, quire16, posit32, quire32, posit_2 or quire_2 output in hexadecimal format.
}
\description{
Converts posit or quire type object to Hexadecimal format.
}
\examples{
a = new ("posit8" , v = -4L)
toHex(a)
}
