% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{toNaR}
\alias{toNaR}
\title{toNaR}
\usage{
toNaR(x)
}
\arguments{
\item{x}{A posit8, posit16, posit32 or posit_2 object.}
}
\description{
Converts an object type posit to NaR (Not a Real).
}
\examples{
a = new ("posit8" , v = 3L)
b = toNaR(a)
}
