% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{toPosit}
\alias{toPosit}
\title{toPosit}
\usage{
toPosit(x)
}
\arguments{
\item{x}{A quire8, quire16, quire32 or quire_2 object.}
}
\value{
posit8, posit16, posit32 or posit_2 object.
}
\description{
Converts quire8, quire16, quire32, quire_2 objects to posit8, posit16, posit32, posit_2 respectively.
}
\examples{
a = new ("quire8")
b = toPosit(a)
a = new ("quire16")
b = toPosit(a)
a = new ("quire32")
b = toPosit(a)
a = new ("quire_2" , y = 6L)
b = toPosit(a)
}
