% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{toPosit16}
\alias{toPosit16}
\title{toPosit16}
\usage{
toPosit16(x)
}
\arguments{
\item{x}{A posit8 or posit32 or  posit_2 object.}
}
\value{
posit16 object.
}
\description{
Converts posit8 or posit32 or posit_2 object to posit16 object.
}
\examples{
a = new ("posit8" , v = 3L)
b = toPosit16(a)
a = new ("posit32" , v = 3L)
b = toPosit16(a)
a = new ("posit_2" , v = 3L , y = 5L)
b = toPosit16(a)
}
