% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{toPosit32}
\alias{toPosit32}
\title{toPosit32}
\usage{
toPosit32(x)
}
\arguments{
\item{x}{A posit8 or posit16 or  posit_2 object.}
}
\value{
posit32 object.
}
\description{
Converts posit8 or posit16 or posit_2 object to posit32 object.
}
\examples{
a = new ("posit8" , v = 3L)
b = toPosit32(a)
a = new ("posit16" , v = 3L)
b = toPosit32(a)
a = new ("posit_2" , v = 3L , y = 5L)
b = toPosit32(a)
}
