% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{toPosit8}
\alias{toPosit8}
\title{toPosit8}
\usage{
toPosit8(x)
}
\arguments{
\item{x}{A posit16 or posit32 or  posit_2 object.}
}
\value{
posit8 object.
}
\description{
Converts posit16 or posit32 or posit_2 object to posit8 object.
}
\examples{
a = new ("posit16" , v = 3L)
b = toPosit8(a)
a = new ("posit32" , v = 3L)
b = toPosit8(a)
a = new ("posit_2" , v = 3L , y = 5L)
b = toPosit8(a)
}
