% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{toPosit_2}
\alias{toPosit_2}
\title{toPosit_2}
\usage{
toPosit_2(x, w)
}
\arguments{
\item{x}{A posit8 or posit16 or  posit32 object.}

\item{w}{An integer value representing the size of posit.}
}
\value{
posit_2 object.
}
\description{
Converts posit8 or posit16 or posit32 object to posit_2 object.
}
\examples{
a = new ("posit8" , v = 3L)
b = toPosit_2(a,5L)
a = new ("posit16" , v = 3L)
b = toPosit_2(a,5L)
a = new ("posit32" , v = 3L)
b = toPosit_2(a,5L)
}
