% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{toRInt}
\alias{toRInt}
\title{toRInt}
\usage{
toRInt(x)
}
\arguments{
\item{x}{A posit8, posit16, posit32 or posit_2 object.}
}
\value{
integer value.
}
\description{
Converts posit type object to integer after rounding.
}
\examples{
a = new ("posit8" , v = 3L)
b = toRInt(a)
a = new ("posit16" , v = 3L)
b = toRInt(a)
a = new ("posit32" , v = 3L)
b = toRInt(a)
}
