% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{type}
\alias{type}
\title{type}
\usage{
type(x)
}
\arguments{
\item{x}{A posit8 or quire8 or posit16 or quire16 or posit32 or quire32 or posit_2 or quire_2 object.}
}
\value{
The type of \code{x}.
}
\description{
Returns the type of posit or quire.
}
\examples{
a = new ("posit8" , v = 3L)
type(a)
a = new ("quire8")
type(a)
a = new ("posit16" , v = 3L)
type(a)
a = new ("quire16")
type(a)
a = new ("posit32" , v = 3L)
type(a)
a = new ("quire32")
type(a)
a = new ("posit_2" , v = 3L , y = 5L)
type(a)
a = new ("quire_2" , y = 5L)
type(a)
}
