% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcript.R
\name{XOR}
\alias{XOR}
\title{XOR}
\usage{
XOR(x, y)
}
\arguments{
\item{x}{A posit8, posit16, posit32 or posit_2 object.}

\item{y}{A posit8, posit16, posit32 or posit_2 object.}
}
\value{
posit8, posit16, posit32 or posit_2 object.
}
\description{
Performs XOR operation on two same posit object types.
}
\examples{
a = new ("posit8" , v = 3L)
b = new ("posit8" , v = 4L)
c = XOR(a,b)
a = new ("posit16" , v = 3L)
b = new ("posit16" , v = 4L)
c = XOR(a,b)
a = new ("posit32" , v = 3L)
b = new ("posit32" , v = 4L)
c = XOR(a,b)
a = new ("posit_2", v =4L, y=5L)
b = new ("posit_2", v =3L, y=5L)
c = XOR(a,b)
}
